pyproject.toml
[project]
name = "python-import-insights"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
]
